pyproject.toml
[project]
name = "ssg"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = ["reflex>=0.8.5", "pyyaml>=6.0.2"]
